app/python/requirement.txt
absl-py==2.0.0
Adafruit-GPIO==1.0.3
Adafruit-PCA9685==1.0.1
Adafruit-PureIO==1.1.11
astroid==3.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
decorator==4.4.2
dill==0.3.7
distlib==0.3.7
docopt==0.6.2
-e git+https://github.com/waveshare/donkeycar@1bbd13f8db4f61c6fe8f734a7608d851e94b7003#egg=donkeycar
filelock==3.12.4
flatbuffers==20181003210633
gast==0.4.0
google-auth==2.23.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.9.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
isort==5.12.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kms==1.0.1.3
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.7.0
moviepy==1.0.3
numpy==1.26.0
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.1.1
picamera==1.13
Pillow==10.0.1
platformdirs==3.11.0
prettytable==3.9.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
RPi.GPIO==0.7.1
rsa==4.9
six==1.16.0
spidev==3.6
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-cpu-aws==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
virtualenv==20.24.5
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.15.0
zipp==3.17.0
